refactor: /admin api 폴더구조 수정
