+++ app/[locale]/admin/page.tsx
@@ -1,4 +1,5 @@
-import { getImportants, getSlides } from '@/apis/admin';
+import { getImportants } from '@/apis/v1/admin/important';
+import { getSlides } from '@/apis/v1/admin/slide';
 import ImportantManagement from '@/app/[locale]/admin/helper/important/ImportantManagement';
 import SlideManagement from '@/app/[locale]/admin/helper/slide/SlideManagement';
 import LoginVisible from '@/components/common/LoginVisible';
